requirements.txt
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cytoolz==1.0.1
eth-hash==0.7.1
eth-typing==5.1.0
eth-utils==5.2.0
idna==3.10
morphys==1.0
py-multibase==1.0.3
pycparser==2.22
pycryptodome==3.21.0
python-baseconv==1.2.2
requests==2.32.3
six==1.17.0
toolz==1.0.0
typing_extensions==4.12.2
urllib3==2.3.0
varint==1.0.2
